Add possibility to access own quizzes' Share screen (#19)
